requirements.txt
<<<<<<< HEAD
Flask==2.3.2
requests==2.31.0
gunicorn==20.1.0
python-dotenv==1.0.0
=======
Flask==2.3.3
openai==1.3.0
nltk==3.8.1
>>>>>>> ba52850ed498ff5ef49dfd5900aec54063f10fd4
